aws_app/app/requirements.txt
streamlit==0.84.1
json5==0.9.5
pillow==8.3.1
glob2==0.7
boto3==1.16.61
botocore==1.19.61